+++ js/widgets/card/cardrenderer.anarchy1.js
@@ -80,6 +80,7 @@ Object.assign(CardRenderer, {
     combatBody += this._drugRow(pnj, "anarchy1", deps);
     combatBody += this._vehicleChipRow(pnj, deps);
     combatBody += this._spiritChipRow(pnj, deps);
+    combatBody += this._spriteCompileRow(pnj, deps);
     combatBody += CyberdeckRenderer.combatArsenal(pnj, "anarchy1"); // râtelier Attaques unifié
     const combatSummary = init != null ? `Init ${init}+${initDice}D6` : "";
     html += this._zoneShell(pnj, "combat", combatBody, combatSummary);